+++ mro/stages/aggregator/count_genes/__init__.py
@@ -42,7 +42,7 @@ def split(args):
         for chunk_start, chunk_len in mol_counter.get_chunks(cr_constants.NUM_MOLECULE_INFO_ENTRIES_PER_CHUNK):
             mol_mem_gb = cr_mol_counter.MoleculeCounter.estimate_mem_gb(chunk_len)
             tot_mem_gb = 4 * (matrix_mem_gb + mol_mem_gb)
-            threads = max(1, round(tot_mem_gb / 8.0))
+            threads = max(1, int(round(tot_mem_gb / 8.0)))
             chunks.append({
                 'chunk_start': str(chunk_start),
                 'chunk_len': str(chunk_len),
